requirements.txt
json5
bs4
GitPython
lxml
requests
py-tlsh
